+++ PiCommon.h
@@ -35,7 +35,7 @@ public:
 	std::string get_current_day();
 	std::string get_yesterday_date();
 
-	float get_temperature();
+	double get_temperature();
 	void uniSleep(int nSleepTime_milisec);
 
 	bool isDateDiff();
